feat: refactor GUI components into separate module
