+++ src/sniffler/gui.py
@@ -7,9 +7,9 @@ from pathlib import Path
 from tkinter import PhotoImage
 
 import customtkinter as ctk
-from tqdm import tqdm
 
 from .collector import Collector
+from .gui_components import AutoHidingScrollableFrame, CTkTqdm
 from .researchers import (
     AudioResearcher,
     BasicResearcher,
@@ -19,35 +19,11 @@ from .researchers import (
     PdfResearcher,
     Researcher,
 )
-from .utils import inherit_signature_from
 
 ICON_PATH = Path(__file__).parent / "assets" / "sniffler.png"
 
 logging.basicConfig(stream=sys.stdout, level=logging.INFO)
 logger = logging.getLogger(__name__)
-
-class AutoHidingScrollableFrame(ctk.CTkScrollableFrame):
-    """
-    Scrollable frame that hides the scrollbar when children element do not exceed the frame.
-    """
-
-    @inherit_signature_from(ctk.CTkScrollableFrame.__init__)
-    def __init__(self, master, **kwargs) -> None:
-        super().__init__(master, **kwargs)
-        self.bind("<Configure>", self._on_configure)
-
-    def _on_configure(self, event=None):
-        self._parent_canvas.configure(scrollregion=self._parent_canvas.bbox("all"))
-        if self._needs_scrollbar():
-            self._scrollbar.grid()
-        else:
-            self._scrollbar.grid_remove()
-
-    def _needs_scrollbar(self):
-        if self._orientation == "vertical":
-            return self._parent_frame.winfo_height() < self.winfo_reqheight()
-        else:
-            return self._parent_frame.winfo_width() < self.winfo_reqwidth()
 
 
 class ChoosePath(ctk.CTkFrame):
@@ -76,25 +52,6 @@ class ChoosePath(ctk.CTkFrame):
         if path:
             self.path = Path(path)
             self.label.configure(text=str(self.path.resolve()))
-
-
-class CTkTqdm(tqdm):
-    def __init__(self, *args, progressbar: ctk.CTkProgressBar | None = None, **kwargs):
-        self.progressbar = progressbar
-        if self.progressbar is not None:
-            self.progressbar.set(0)
-        super().__init__(*args, **kwargs)
-
-    def display(self, *args, **kwargs):
-        pass  # Prevent console output
-
-    def refresh(self, *args, **kwargs):
-        pass  # Prevent console output
-
-    def update(self, n=1) -> None:
-        super().update(n)
-        if self.progressbar and self.total:
-            self.progressbar.set(self.n / self.total)
 
 
 class AppController:
